Improve unit test summaries
