+++ tests/v2_x/test_event_mechanics.py
@@ -38,7 +38,7 @@ start_main_flow_event = InternalEvent(name="StartFlow", arguments={"flow_id": "m
 
 
 def test_send_umim_action_event():
-    """Test to start an UMIM event"""
+    """Test to send an UMIM event."""
 
     content = """
     flow main
@@ -58,7 +58,7 @@ def test_send_umim_action_event():
 
 
 def test_match_umim_action_event():
-    """Test to match an UMIM event"""
+    """Test to match an UMIM event."""
 
     content = """
     flow main
@@ -120,7 +120,7 @@ def test_start_action():
 
 
 def test_start_match_action_on_action_parameter():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and match an UMIM action based on action parameters."""
 
     content = """
     flow main
@@ -161,7 +161,7 @@ def test_start_match_action_on_action_parameter():
 
 
 def test_start_mismatch_action_on_action_parameter():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and mismatch an UMIM action based on action parameters."""
 
     content = """
     flow main
@@ -194,7 +194,7 @@ def test_start_mismatch_action_on_action_parameter():
 
 
 def test_start_match_action_on_event_parameter():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and match an UMIM action based on action parameters."""
 
     content = """
     flow main
@@ -235,7 +235,7 @@ def test_start_match_action_on_event_parameter():
 
 
 def test_start_mismatch_action_on_event_parameter():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and mismatch an UMIM action based on event parameters."""
 
     content = """
     flow main
@@ -268,7 +268,7 @@ def test_start_mismatch_action_on_event_parameter():
 
 
 def test_start_match_action_with_reference():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and match an UMIM action based on action reference."""
 
     content = """
     flow main
@@ -309,7 +309,7 @@ def test_start_match_action_with_reference():
 
 
 def test_await_action():
-    """Test to await an UMIM action"""
+    """Test to await an UMIM action."""
 
     content = """
     flow main
@@ -349,7 +349,7 @@ def test_await_action():
 
 
 def test_implicit_action_state_update():
-    """Test the action state update"""
+    """Test the implicit action state update."""
 
     content = """
     flow main
@@ -383,8 +383,8 @@ def test_implicit_action_state_update():
     assert state.actions[action_uid].status == ActionStatus.FINISHED
 
 
-def test_start_a_flow():
-    """Test the start of a child flow with full event notation"""
+def test_start_a_flow_expanded_notation():
+    """Test the start of a child flow with expanded notation."""
 
     content = """
     flow a
@@ -392,8 +392,8 @@ def test_start_a_flow():
 
     flow main
       # start a
-      send StartFlow(flow_id="a")
-      match FlowStarted(flow_id="a")
+      send StartFlow(flow_id="a", flow_instance_uid="01234")
+      match FlowStarted(flow_id="a", flow_instance_uid="01234")
     """
 
     state = run_to_completion(_init_state(content), start_main_flow_event)
@@ -411,8 +411,8 @@ def test_start_a_flow():
     )
 
 
-def test_start_a_flow_compact_notation():
-    """Test the start of a child flow using 'start' notation"""
+def test_start_a_flow_short_notation():
+    """Test the start of a child flow using short notation."""
 
     content = """
     flow a
@@ -438,7 +438,7 @@ def test_start_a_flow_compact_notation():
 
 
 def test_start_match_flow_with_reference():
-    """Test to start and match an UMIM action based on action parameters"""
+    """Test to start and match end of a flow based on flow reference."""
 
     content = """
     flow bot say hello
@@ -481,8 +481,77 @@ def test_start_match_flow_with_reference():
     )
 
 
+def test_user_action_reference():
+    """Test user action event matching based on references."""
+
+    content = """
+    flow main
+      match UtteranceUserAction.Started() as $event_ref
+      start UtteranceBotAction(script="Started user action: {$event_ref.action.name}")
+      match $event_ref.action.Finished(final_transcript="End")
+      start UtteranceBotAction(script="Success")
+    """
+
+    config = _init_state(content)
+    state = run_to_completion(config, start_main_flow_event)
+    assert is_data_in_events(
+        state.outgoing_events,
+        [],
+    )
+    state = run_to_completion(
+        state,
+        {
+            "type": "UtteranceUserActionStarted",
+            "uid": "d4a265bb-4a27-4d28-8ca5-80cc73dc4707",
+            "event_created_at": "2023-09-12T13:01:16.334940+00:00",
+            "source_uid": "umim_tui_app",
+            "action_uid": "cc63b1a0-5703-4e80-b66b-2734c13abcf3",
+            "action_info_modality": "user_speech",
+        },
+    )
+    assert is_data_in_events(
+        state.outgoing_events,
+        [
+            {
+                "type": "StartUtteranceBotAction",
+                "script": "Started user action: UtteranceUserAction",
+            },
+        ],
+    )
+    state = run_to_completion(
+        state,
+        {
+            "type": "UtteranceUserActionFinished",
+            "uid": "d4a265bb-4a27-4d28-8ca5-80cc73dc4707",
+            "event_created_at": "2023-09-12T13:01:16.334940+00:00",
+            "source_uid": "umim_tui_app",
+            "action_uid": "cc63b1a0-5703-4e80-b66b-2734c13abcf3",
+            "final_transcript": "End",
+            "is_success": True,
+            "action_info_modality": "user_speech",
+            "action_info_modality_policy": "replace",
+            "action_finished_at": "2023-09-12T13:01:16.334954+00:00",
+        },
+    )
+    assert is_data_in_events(
+        state.outgoing_events,
+        [
+            {
+                "type": "StartUtteranceBotAction",
+                "script": "Success",
+            },
+            {
+                "type": "StopUtteranceBotAction",
+            },
+            {
+                "type": "StopUtteranceBotAction",
+            },
+        ],
+    )
+
+
 def test_await_a_flow():
-    """Test await a child flow"""
+    """Test await a child flow."""
 
     content = """
     flow a
@@ -490,9 +559,9 @@ def test_await_a_flow():
 
     flow main
       # await a
-      send StartFlow(flow_id="a")
-      match FlowStarted(flow_id="a")
-      match FlowFinished(flow_id="a")
+      send StartFlow(flow_id="a", flow_instance_uid="01234")
+      match FlowStarted(flow_id="a", flow_instance_uid="01234")
+      match FlowFinished(flow_id="a", flow_instance_uid="01234")
       start UtteranceBotAction(script="Flow a finished")
     """
 
@@ -518,8 +587,8 @@ def test_await_a_flow():
     )
 
 
-def test_await_a_flow_compact_notation():
-    """Test await a child flow with compact notation 'await'"""
+def test_await_a_flow_short_notation():
+    """Test await a child flow with short notation."""
 
     content = """
     flow a
@@ -552,8 +621,43 @@ def test_await_a_flow_compact_notation():
     )
 
 
+def test_await_multimodal_action():
+    """Test await statement with multimodal actions."""
+
+    content = """
+    flow bot say $text
+      await UtteranceBotAction(script=$text) as $action
+
+    flow bot gesture $gesture
+      await GestureBotAction(gesture=$gesture) as $action
+
+    flow bot express $text
+      await bot say $text
+
+    flow main
+        start bot express "Hi"
+        start bot gesture "Wave"
+        match UtteranceUserAction().Finished()
+    """
+
+    state = run_to_completion(_init_state(content), start_main_flow_event)
+    assert is_data_in_events(
+        state.outgoing_events,
+        [
+            {
+                "type": "StartUtteranceBotAction",
+                "script": "Hi",
+            },
+            {
+                "type": "StartGestureBotAction",
+                "gesture": "Wave",
+            },
+        ],
+    )
+
+
 def test_start_child_flow_two_times():
-    """Test start a child flow two times"""
+    """Test start a child flow two times."""
 
     content = """
     flow a
@@ -580,8 +684,8 @@ def test_start_child_flow_two_times():
     )
 
 
-def test_event_simple_parameter_match():
-    """Test start a child flow two times"""
+def test_simple_event_parameter_match():
+    """Test a simple event parameter match."""
 
     content = """
     flow main
@@ -662,8 +766,8 @@ def test_event_simple_parameter_match():
     )
 
 
-def test_event_dict_parameter_match():
-    """Test start a child flow two times"""
+def test_dict_event_parameter_match():
+    """Test event match with dictionary parameters."""
 
     content = """
     flow main
@@ -744,8 +848,8 @@ def test_event_dict_parameter_match():
     )
 
 
-def test_event_set_parameter_match():
-    """Test start a child flow two times"""
+def test_set_event_parameter_match():
+    """Test event match with set parameters."""
 
     content = """
     flow main
@@ -844,8 +948,8 @@ def test_event_set_parameter_match():
     )
 
 
-def test_event_list_parameter_match():
-    """Test start a child flow two times"""
+def test_list_event_parameter_match():
+    """Test event match with list parameters."""
 
     content = """
     flow main
@@ -966,8 +1070,8 @@ def test_event_list_parameter_match():
     )
 
 
-def test_event_custom_regex_parameter_match():
-    """Test more complex regex parameters."""
+def test_custom_regex_event_parameter_match():
+    """Test more complex regex event parameters."""
 
     content = """
     flow main
@@ -1025,8 +1129,8 @@ def test_event_custom_regex_parameter_match():
     )
 
 
-def test_event_corner_cases_regex_parameter_match():
-    """Test corner cases for regex matches"""
+def test_corner_cases_regex_event_parameter_matching():
+    """Test corner cases for regex matches."""
 
     content = """
     flow main
@@ -1070,6 +1174,7 @@ def test_event_corner_cases_regex_parameter_match():
 
 
 def test_action_event_requeuing():
+    """Test queuing of action events."""
     config = RailsConfig.from_content(
         colang_content="""
         flow main
@@ -1092,8 +1197,8 @@ def test_action_event_requeuing():
     chat << "started\nsuccess"
 
 
-def test_update_context():
-    """Test to update the context."""
+def test_context_update():
+    """Test to update the context through events."""
 
     content = """
     flow main
@@ -1170,8 +1275,8 @@ def test_update_context():
     )
 
 
-def test_match_internal_event():
-    """Test to update the context."""
+def test_match_hierarchy_of_internal_events():
+    """Test internal event matching for a typical flow hierarchy."""
 
     content = """
     flow _bot_say $text
