+++ tests/v2_x/test_slide_mechanics.py
@@ -36,7 +36,7 @@ start_main_flow_event = InternalEvent(name="StartFlow", arguments={"flow_id": "m
 
 
 def test_while_loop_mechanic():
-    """"""
+    """Test the while loop statement mechanic."""
 
     content = """
     flow main
@@ -82,114 +82,8 @@ def test_while_loop_mechanic():
     )
 
 
-def test_await_multimodal_action():
-    """"""
-
-    content = """
-    flow bot say $text
-      await UtteranceBotAction(script=$text) as $action
-
-    flow bot gesture $gesture
-      await GestureBotAction(gesture=$gesture) as $action
-
-    flow bot express $text
-      await bot say $text
-
-    flow main
-        start bot express "Hi"
-        start bot gesture "Wave"
-        match UtteranceUserAction().Finished()
-    """
-
-    state = run_to_completion(_init_state(content), start_main_flow_event)
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "Hi",
-            },
-            {
-                "type": "StartGestureBotAction",
-                "gesture": "Wave",
-            },
-        ],
-    )
-
-
-def test_activate_and_grouping():
-    """"""
-
-    content = """
-    flow a
-      start UtteranceBotAction(script="A")
-      match UtteranceUserAction().Finished(final_transcript="a")
-
-    flow b
-      start UtteranceBotAction(script="B")
-      match UtteranceUserAction().Finished(final_transcript="b")
-
-    flow main
-        activate a and b
-        match UtteranceUserAction().Finished(final_transcript="end")
-    """
-
-    state = run_to_completion(_init_state(content), start_main_flow_event)
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "A",
-            },
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "B",
-            },
-        ],
-    )
-    state = run_to_completion(
-        state,
-        {
-            "type": "UtteranceUserActionFinished",
-            "final_transcript": "a",
-        },
-    )
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StopUtteranceBotAction",
-            },
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "A",
-            },
-        ],
-    )
-    state = run_to_completion(
-        state,
-        {
-            "type": "UtteranceUserActionFinished",
-            "final_transcript": "b",
-        },
-    )
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StopUtteranceBotAction",
-            },
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "B",
-            },
-        ],
-    )
-
-
 def test_if_branching_mechanic():
-    """"""
+    """Test if branching statement mechanism."""
 
     content = """
     flow main
@@ -255,7 +149,7 @@ def test_if_branching_mechanic():
 
 
 def test_event_reference_member_access():
-    """"""
+    """Test accessing a event reference member."""
 
     content = """
     flow main
@@ -291,7 +185,7 @@ def test_event_reference_member_access():
 
 
 def test_action_reference_member_access():
-    """"""
+    """Test accessing a action reference member."""
 
     content = """
     flow main
@@ -323,7 +217,7 @@ def test_action_reference_member_access():
 
 
 def test_flow_references_member_access():
-    """"""
+    """Test accessing a flow reference member."""
 
     content = """
     flow bot say $text
@@ -358,7 +252,7 @@ def test_flow_references_member_access():
 
 
 def test_expressions_in_strings():
-    """"""
+    """Test string expression evaluation."""
 
     content = """
     flow main
@@ -398,7 +292,7 @@ def test_expressions_in_strings():
 
 
 def test_flow_return_values():
-    """"""
+    """Test flow return value handling."""
 
     content = """
     flow a
@@ -435,7 +329,7 @@ def test_flow_return_values():
 
 
 def test_break_continue_statement_a():
-    """"""
+    """Test break and continue statements within while loop."""
 
     content = """
     flow main
@@ -505,7 +399,7 @@ def test_break_continue_statement_a():
 
 
 def test_break_continue_statement_b():
-    """"""
+    """Test break and continue statements within while loop."""
 
     content = """
     flow main
@@ -551,7 +445,7 @@ def test_break_continue_statement_b():
 
 
 def test_when_or_core_mechanics():
-    """"""
+    """Test when / or when statement mechanics."""
 
     content = """
     flow user said $transcript
@@ -634,7 +528,7 @@ def test_when_or_core_mechanics():
 
 
 def test_when_or_bot_action_mechanics():
-    """"""
+    """Test when / or when statement mechanics with actions."""
 
     content = """
     flow main
@@ -706,7 +600,7 @@ def test_when_or_bot_action_mechanics():
 
 
 def test_when_or_group_mechanics():
-    """"""
+    """Test when / or when statement mechanics with or-grouping."""
 
     content = """
     flow user said $transcript
@@ -800,7 +694,7 @@ def test_when_or_group_mechanics():
 
 
 def test_when_or_competing_events_mechanics():
-    """"""
+    """Test when / or when statement mechanics with events."""
 
     content = """
     flow user said something
@@ -886,7 +780,7 @@ def test_when_or_competing_events_mechanics():
 
 
 def test_when_or_with_references():
-    """"""
+    """Test when / or when statement mechanics with references."""
 
     content = """
     flow user said something
@@ -923,7 +817,7 @@ def test_when_or_with_references():
 
 
 def test_inside_when_failure_handling():
-    """"""
+    """Test when / or when statement failure handling mechanics."""
 
     content = """
     flow a
@@ -974,7 +868,7 @@ def test_inside_when_failure_handling():
 
 
 def test_abort_flow():
-    """"""
+    """Test abort keyword mechanics."""
 
     content = """
     flow a
@@ -1016,7 +910,7 @@ def test_abort_flow():
 
 
 def test_global_statement():
-    """"""
+    """Test global variables."""
 
     content = """
     flow a
@@ -1061,8 +955,8 @@ def test_global_statement():
     )
 
 
-def test_public_flow_attributes():
-    """"""
+def test_out_flow_variables():
+    """Test the out variables flow mechanics."""
 
     content = """
     flow a -> $result_1 = 1, $result_2 = 2
