+++ tests/v2_x/test_various_mechanics.py
@@ -36,7 +36,7 @@ start_main_flow_event = InternalEvent(name="StartFlow", arguments={"flow_id": "m
 
 
 def test_multi_flow_level_member_access():
-    """"""
+    """Test access of child of child flow attributes."""
 
     content = """
     flow user said $transcript
@@ -79,8 +79,8 @@ def test_multi_flow_level_member_access():
     )
 
 
-def test_FlowStart_event_fallback():
-    """"""
+def test_flow_start_event_fallback():
+    """Test the flow start fallback case."""
 
     content = """
     flow a
@@ -118,7 +118,7 @@ def test_FlowStart_event_fallback():
 
 
 def test_multi_level_member_match_from_reference():
-    """"""
+    """Test matching based on reference attributes."""
 
     content = """
     flow a
@@ -158,8 +158,8 @@ def test_multi_level_member_match_from_reference():
     )
 
 
-def test_flow_deactivation_on_parent_flow_finished():
-    """"""
+def test_flow_end_on_parent_flow_finished():
+    """Check that a flow gets stopped when its parent flow has ended."""
 
     content = """
     flow a
@@ -198,7 +198,7 @@ def test_flow_deactivation_on_parent_flow_finished():
 
 
 def test_event_action_wrapper_abstraction():
-    """"""
+    """Test a common action abstraction use case."""
 
     content = """
     flow user said $text
@@ -276,104 +276,6 @@ def test_event_action_wrapper_abstraction():
             }
         ],
     )
-
-
-# TODO: Check if we really need this
-# def test_start_sibling_flow_mechanism():
-#     """"""
-
-#     content = """
-#     flow a
-#       match UtteranceUserAction.Finished(final_transcript="1")
-#       send StartSiblingFlow(flow_id="a")
-#       start UtteranceBotAction(script="A")
-#       match UtteranceUserAction.Finished(final_transcript="2")
-#       start UtteranceBotAction(script="B")
-
-#     flow main
-#       activate a
-#       match UtteranceUserAction.Finished(final_transcript="End")
-#     """
-
-#     config = _init_state(content)
-#     state = run_to_completion(config, start_main_flow_event)
-#     assert is_data_in_events(
-#         state.outgoing_events,
-#         [],
-#     )
-#     state = run_to_completion(
-#         state,
-#         {
-#             "type": "UtteranceUserActionFinished",
-#             "final_transcript": "1",
-#         },
-#     )
-#     assert is_data_in_events(
-#         state.outgoing_events,
-#         [
-#             {
-#                 "type": "StartUtteranceBotAction",
-#                 "script": "A",
-#             },
-#         ],
-#     )
-#     state = run_to_completion(
-#         state,
-#         {
-#             "type": "UtteranceUserActionFinished",
-#             "final_transcript": "1",
-#         },
-#     )
-#     assert is_data_in_events(
-#         state.outgoing_events,
-#         [
-#             {
-#                 "type": "StartUtteranceBotAction",
-#                 "script": "A",
-#             },
-#         ],
-#     )
-#     state = run_to_completion(
-#         state,
-#         {
-#             "type": "UtteranceUserActionFinished",
-#             "final_transcript": "2",
-#         },
-#     )
-#     assert is_data_in_events(
-#         state.outgoing_events,
-#         [
-#             {
-#                 "type": "StartUtteranceBotAction",
-#                 "script": "B",
-#             },
-#             {
-#                 "type": "StopUtteranceBotAction",
-#             },
-#             {
-#                 "type": "StopUtteranceBotAction",
-#             },
-#             {
-#                 "type": "StopUtteranceBotAction",
-#             },
-#         ],
-#     )
-#     state = run_to_completion(
-#         state,
-#         {
-#             "type": "UtteranceUserActionFinished",
-#             "final_transcript": "1",
-#         },
-#     )
-#     assert is_data_in_events(
-#         state.outgoing_events,
-#         [
-#             {
-#                 "type": "StartUtteranceBotAction",
-#                 "script": "A",
-#             },
-#         ],
-#     )
 
 
 # def test_stop_flow():
@@ -461,77 +363,8 @@ def test_event_action_wrapper_abstraction():
 #     )
 
 
-def test_user_action_reference():
-    """"""
-
-    content = """
-    flow main
-      match UtteranceUserAction.Started() as $event_ref
-      start UtteranceBotAction(script="Started user action: {$event_ref.action.name}")
-      match $event_ref.action.Finished(final_transcript="End")
-      start UtteranceBotAction(script="Success")
-    """
-
-    config = _init_state(content)
-    state = run_to_completion(config, start_main_flow_event)
-    assert is_data_in_events(
-        state.outgoing_events,
-        [],
-    )
-    state = run_to_completion(
-        state,
-        {
-            "type": "UtteranceUserActionStarted",
-            "uid": "d4a265bb-4a27-4d28-8ca5-80cc73dc4707",
-            "event_created_at": "2023-09-12T13:01:16.334940+00:00",
-            "source_uid": "umim_tui_app",
-            "action_uid": "cc63b1a0-5703-4e80-b66b-2734c13abcf3",
-            "action_info_modality": "user_speech",
-        },
-    )
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "Started user action: UtteranceUserAction",
-            },
-        ],
-    )
-    state = run_to_completion(
-        state,
-        {
-            "type": "UtteranceUserActionFinished",
-            "uid": "d4a265bb-4a27-4d28-8ca5-80cc73dc4707",
-            "event_created_at": "2023-09-12T13:01:16.334940+00:00",
-            "source_uid": "umim_tui_app",
-            "action_uid": "cc63b1a0-5703-4e80-b66b-2734c13abcf3",
-            "final_transcript": "End",
-            "is_success": True,
-            "action_info_modality": "user_speech",
-            "action_info_modality_policy": "replace",
-            "action_finished_at": "2023-09-12T13:01:16.334954+00:00",
-        },
-    )
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "Success",
-            },
-            {
-                "type": "StopUtteranceBotAction",
-            },
-            {
-                "type": "StopUtteranceBotAction",
-            },
-        ],
-    )
-
-
 def test_stop_bot_action():
-    """"""
+    """Test stopping an action."""
 
     content = """
     flow main
@@ -565,44 +398,8 @@ def test_stop_bot_action():
     )
 
 
-# TODO: Double-check if the behavior of an or-group makes sense in combination with separated interaction loops
-def test_independent_flow_loop_mechanics():
-    """"""
-
-    content = """
-    @loop("NEW")
-    flow bot say $script
-      await UtteranceBotAction(script=$script)
-
-    flow main
-      start bot say "Hi" or bot say "Hello"
-    """
-
-    config = _init_state(content)
-    state = run_to_completion(config, start_main_flow_event)
-    assert is_data_in_events(
-        state.outgoing_events,
-        [
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "Hi",
-            },
-            {
-                "type": "StartUtteranceBotAction",
-                "script": "Hello",
-            },
-            {
-                "type": "StopUtteranceBotAction",
-            },
-            {
-                "type": "StopUtteranceBotAction",
-            },
-        ],
-    )
-
-
 def test_list_parameters():
-    """"""
+    """Test list flow parameters."""
 
     content = """
     flow bot say $scripts
@@ -629,7 +426,7 @@ def test_list_parameters():
 
 
 def test_dict_parameters():
-    """"""
+    """Test dict flow parameters."""
 
     content = """
     flow bot say $scripts
@@ -656,19 +453,16 @@ def test_dict_parameters():
 
 
 def test_mixed_multimodal_group_actions():
-    """"""
+    """Testing common used case of mixed multimodal flows."""
 
     content = """
     flow bot say $text
-      # meta: exclude from llm
       await UtteranceBotAction(script=$text) as $action
 
     flow bot gesture $gesture
-      # meta: exclude from llm
       await GestureBotAction(gesture=$gesture) as $action
 
     flow bot express $text
-      # meta: exclude from llm
       await bot say $text
 
     flow bot express feeling well
@@ -680,7 +474,6 @@ def test_mixed_multimodal_group_actions():
         and (bot gesture "Thumbs down" or bot gesture "Sad face")
 
     flow main
-      #bot say "One" and (bot gesture "Two" or bot gesture "Three")
       bot express feeling well
         or bot express feeling bad
       match NeverComingEvent()
@@ -726,7 +519,7 @@ def test_mixed_multimodal_group_actions():
 
 
 def test_iternal_unhandled_event():
-    """"""
+    """Test unhandled event handling."""
 
     content = """
     flow undefined flows
@@ -774,7 +567,7 @@ def test_iternal_unhandled_event():
 
 
 def test_references_in_groups():
-    """"""
+    """Test use of references in groups in child flow."""
 
     content = """
     flow bot say $script
@@ -786,7 +579,7 @@ def test_references_in_groups():
 
     flow main
       bot greets as $ref
-      bot say $ref.context.ref.context.action.context.script
+      bot say $ref.ref.context.action.context.script
       match NeverComingEvent()
     """
 
@@ -821,7 +614,7 @@ def test_references_in_groups():
 
 
 def test_regular_expressions_action_parameters():
-    """"""
+    """Test use of regex in action parameters."""
 
     content = """
     flow main
@@ -854,7 +647,7 @@ def test_regular_expressions_action_parameters():
 
 
 def test_regular_expressions_flow_parameters():
-    """"""
+    """Test use of regex in flow parameters."""
 
     content = """
     flow user said $transcript
@@ -890,7 +683,7 @@ def test_regular_expressions_flow_parameters():
 
 
 def test_expr_func_search():
-    """"""
+    """Test Colang function search."""
 
     content = """
     flow main
@@ -913,7 +706,7 @@ def test_expr_func_search():
 
 
 def test_generate_value_with_NLD():
-    """"""
+    """Test NLD instructions."""
 
     content = """
     flow main
